Fix multiple input in same grid block
Before this commit the mouse input was taken to set X or O in a grid
block without checking if that block was previously marked. Thus a
player can mark a block in place of the other player. For example, if a
block was previously marked by a player with X, the other player could
mark that same block with O.

This commit will fix this issue. It just needed to check whether the
block clicked on by the mouse is blank or filled.